Graphql structure for Projects
