--- client/pages/Projects/GET_PROJECTS.ts
@@ -1,33 +1,0 @@
-import gql from 'graphql-tag'
-
-export default gql`
-  query($customerKey: String, $sortBy: String) {
-    projects(customerKey: $customerKey, sortBy: $sortBy) {
-      id
-      key
-      name
-      description
-      webLink
-      icon
-      externalSystemURL
-      customerKey
-      customer {
-        key
-        name
-      }
-      labels {
-        name
-        description
-        color
-        icon
-      }
-      inactive
-    }
-    outlookCategories {
-      id
-      key
-      displayName
-      color
-    }
-  }
-`
