--- client/pages/Projects/ProjectDetails/CREATE_OUTLOOK_CATEGORY.ts
@@ -1,13 +1,0 @@
-import gql from 'graphql-tag'
-
-export const CREATE_OUTLOOK_CATEGORY = gql`
-  mutation($category: String!) {
-    result: createOutlookCategory(category: $category) {
-      data
-      success
-      error {
-        message
-      }
-    }
-  }
-`
