+++ client/pages/Projects/ProjectForm/index.tsx
@@ -8,8 +8,8 @@ import React, { useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { IFormValidation } from 'types'
 import { isBlank } from 'underscore.string'
+import graphql from '../graphql'
 import styles from './CreateProjectForm.module.scss'
-import CREATE_OR_UPDATE_PROJECT, { ICreateOrUpdateProjectVariables } from './CREATE_OR_UPDATE_PROJECT'
 import { IProjectFormProps, ProjectModel } from './types'
 
 
@@ -19,7 +19,7 @@ export const ProjectForm = ({ edit, onSubmitted, nameLength = [2] }: IProjectFor
     const [validation, setValidation] = useState<IFormValidation>({ errors: {}, invalid: true })
     const [message, setMessage] = useMessage()
     const [model, setModel] = useState<ProjectModel>(new ProjectModel(edit))
-    const [createOrUpdateProject, { loading }] = useMutation<any, ICreateOrUpdateProjectVariables>(CREATE_OR_UPDATE_PROJECT)
+    const [createOrUpdateProject, { loading }] = useMutation<any, any>(graphql.mutation.createOutlookCategory)
 
     /**
      * Update model
